pyproject.toml
[tool.poetry]
name = "rmh"
version = "1"
description = "RoomMateHub"
authors = ["Tech'Titans"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-rest-framework = "^0.1.0"
django = "^5.0.3"
cors = "^1.0.1"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
